+++ sdl_example/SDLSignalDisplayProcessor.cpp
@@ -45,7 +45,7 @@ SDLSignalDisplayProcessor::SDLSignalDisplayProcessor(LFPBuffer *buffer, std::str
 	Log("Constructor start");
 
     SDL_SetRenderDrawColor(renderer_, 255, 0, 0, 255);
-    SDL_RenderDrawLine(renderer_, 1, SHIFT/plot_scale, window_width, SHIFT/plot_scale);
+    SDL_RenderDrawLine(renderer_, 1, int(SHIFT/plot_scale), (int)window_width, int(SHIFT/plot_scale));
     prev_vals_ = new int[buffer->CHANNEL_NUM];
 
     // check if lfp display channels is a subset of read channels
@@ -133,7 +133,7 @@ void SDLSignalDisplayProcessor::process(){
 int SDLSignalDisplayProcessor::transform_to_y_coord(int voltage){
     // scale for plotting
     int val = voltage + SHIFT;
-    val = val > 0 ? val / plot_scale : 1;
+    val = val > 0 ? int(val / plot_scale) : 1;
     val = (unsigned int)val < SCREEN_HEIGHT ? val : SCREEN_HEIGHT;
     
     return val;
